update graph eliminations: c++14 + can cope with domain sizes defined over sets of nodes different from those of the graph

+++ src/agrum/graphs/simplicialSet.inl
@@ -33,19 +33,13 @@
 
 namespace gum {
 
-  /* ===========================================================================
-   */
-  /* ===========================================================================
-   */
-  /* ===  CLASS FOR RETRIEVING SIMPLICIAL, ALMOST AND QUASI SIMPLICIAL NODES ===
-   */
-  /* ===========================================================================
-   */
-  /* ===========================================================================
-   */
+  /* =========================================================================== */
+  /* =========================================================================== */
+  /* ===  CLASS FOR RETRIEVING SIMPLICIAL, ALMOST AND QUASI SIMPLICIAL NODES === */
+  /* =========================================================================== */
+  /* =========================================================================== */
 
   /// indicates whether a given node is a simplicial node
-
   INLINE
   bool SimplicialSet::isSimplicial( const NodeId id ) {
     // update the list to which the node belongs if needed
@@ -55,8 +49,8 @@ namespace gum {
     return __simplicial_nodes.contains( id );
   }
 
+  
   /// gets a simplicial node
-
   INLINE
   NodeId SimplicialSet::bestSimplicialNode() {
     if ( !hasSimplicialNode() ) {
@@ -66,8 +60,8 @@ namespace gum {
     return __simplicial_nodes.top();
   }
 
+  
   /// gets an almost simplicial node
-
   INLINE
   NodeId SimplicialSet::bestAlmostSimplicialNode() {
     if ( !hasAlmostSimplicialNode() ) {
@@ -77,8 +71,8 @@ namespace gum {
     return __almost_simplicial_nodes.top();
   }
 
+  
   /// gets a quasi simplicial node
-
   INLINE
   NodeId SimplicialSet::bestQuasiSimplicialNode() {
     if ( !hasQuasiSimplicialNode() ) {
@@ -88,52 +82,58 @@ namespace gum {
     return __quasi_simplicial_nodes.top();
   }
 
+  
   /// put all the nodes in their appropriate list
-
   INLINE
   void SimplicialSet::__updateAllNodes() {
     // check if a node can enter the simplicial list
-    for ( auto iter =
-              __changed_status.beginSafe();  // safe iterator needed here
+    for ( auto iter = __changed_status.beginSafe();  // safe iterator needed here
           iter != __changed_status.endSafe();
           ++iter ) {
       __updateList( *iter );
     }
   }
+  
 
   /// returns all the simplicial nodes
-
   INLINE
-  const PriorityQueue<NodeId, float>& SimplicialSet::allSimplicialNodes() {
+  const PriorityQueue<NodeId, double>& SimplicialSet::allSimplicialNodes() {
     __updateAllNodes();
     return __simplicial_nodes;
   }
 
+  
   /// returns all the almost simplicial nodes
-
   INLINE
-  const PriorityQueue<NodeId, float>&
+  const PriorityQueue<NodeId, double>&
   SimplicialSet::allAlmostSimplicialNodes() {
     __updateAllNodes();
     return __almost_simplicial_nodes;
   }
 
+  
   /// returns all the quasi simplicial nodes
-
   INLINE
-  const PriorityQueue<NodeId, float>& SimplicialSet::allQuasiSimplicialNodes() {
+  const PriorityQueue<NodeId, double>& SimplicialSet::allQuasiSimplicialNodes() {
     __updateAllNodes();
     return __quasi_simplicial_nodes;
   }
+  
 
   /// sets/unset the fill-ins storage in the standard triangulation procedure
-
   INLINE
-  void SimplicialSet::setFillIns( bool b ) { __we_want_fill_ins = b; }
+  void SimplicialSet::setFillIns( bool b ) {
+    __we_want_fill_ins = b;
+  }
 
+
+  /// returns the set of fill-ins
   INLINE
-  const EdgeSet& SimplicialSet::fillIns() const { return __fill_ins_list; }
+  const EdgeSet& SimplicialSet::fillIns() const {
+    return __fill_ins_list;
+  }
 
+  
 } /* namespace gum */
 
 #endif /* DOXYGEN_SHOULD_SKIP_THIS */
